requirements.txt
google-genai
duckdb
ipykernel
google
python-dotenv
psycopg2-binary